+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- Added new tour for all monitoring locations of interest
+- Changed the popup so only the one shows at a time